requirements.txt
Flask==1.0.2
beautifulsoup4==4.9.1
xlsxwriter==1.2.8
google-cloud-storage
